backend/requirements.txt
# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Base de datos
pony>=0.7.19
psycopg2-binary>=2.9.9

# Autenticación y seguridad
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2
python-decouple>=3.8

# Validación (incluido en FastAPI, explícito para compatibilidad)
pydantic>=2.5.0

# Reportes PDF
reportlab>=4.0.0
